Trenger ikke å konvertere til Int nå som sluttbeløpet allerede er Int

+++ src/main/kotlin/no/nav/familie/ef/sak/behandlingsflyt/steg/BeregnYtelseSteg.kt
@@ -318,9 +318,11 @@ class BeregnYtelseSteg(private val tilkjentYtelseService: TilkjentYtelseService,
 
     private fun lagBeløpsperioderForInnvilgelseBarnetilsyn(vedtak: InnvilgelseBarnetilsyn,
                                                            saksbehandling: Saksbehandling) =
-            beregningBarnetilsynService.beregnYtelseBarnetilsyn(vedtak.perioder, vedtak.perioderKontantstøtte, vedtak.tilleggsstønad.perioder)
+            beregningBarnetilsynService.beregnYtelseBarnetilsyn(vedtak.perioder,
+                                                                vedtak.perioderKontantstøtte,
+                                                                vedtak.tilleggsstønad.perioder)
                     .map {
-                        AndelTilkjentYtelse(beløp = it.beløp.toInt(),
+                        AndelTilkjentYtelse(beløp = it.beløp,
                                             stønadFom = it.periode.fradato,
                                             stønadTom = it.periode.tildato,
                                             kildeBehandlingId = saksbehandling.id,
